packages/web-app-files/tests/unit/helpers/resource/resource.spec.ts
import {
  extractExtensionFromFile,
  extractNameWithoutExtension,
  Resource
} from '../../../../src/helpers/resource'

const resourceWithoutExtension = {
  name: 'file'
}
const resourceNameWithExtension = {
  name: 'file.txt',
  extension: 'txt'
}
const resourceNameWithExtensionAndDots = {
  name: 'file.dot.txt',
  extension: 'txt'
}

describe('filterResources', () => {
  describe('extractNameWithoutExtension', () => {
    it('should return resource name when there is no extension', () => {
      expect(extractNameWithoutExtension(resourceWithoutExtension as Resource)).toEqual(
        resourceWithoutExtension.name
      )
    })
    it('should return resource name without extension when there is an extension', () => {
      expect(extractNameWithoutExtension(resourceNameWithExtension as Resource)).toEqual('file')
    })
    it('should return resource name with dots without extension when there is an extension', () => {
      expect(extractNameWithoutExtension(resourceNameWithExtensionAndDots as Resource)).toEqual(
        'file.dot'
      )
    })
  })
  describe('extractExtensionFromFile', () => {
    it('should return simple file extension', () => {
      expect(extractExtensionFromFile('test.png')).toEqual('png')
    })
    it('should return complex file extension', () => {
      expect(extractExtensionFromFile('test.tar.gz')).toEqual('tar.gz')
    })
    it('should return unknown file extension', () => {
      expect(extractExtensionFromFile('test.unknown')).toEqual('unknown')
    })
    it('should return no file extension', () => {
      expect(extractExtensionFromFile('test')).toEqual('')
    })
  })
})
